backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pdfplumber==0.11.4
python-docx==1.1.2
groq>=0.13.0
sentence-transformers==3.1.1
sqlalchemy==2.0.35
pg8000==1.31.2
httpx==0.28.1
